fix: bind PIN reset to the current target user
