+++ docs/API.md
@@ -214,7 +214,9 @@ PIN 設定後、正式な 24 時間セッションへ切り替え、端末単位
 
 仕様:
 
-- 登録済みメールアドレスのみ受け付け、未登録メールアドレスは `400` を返します。
+- remembered user がいる端末からのみ利用できます。対象ユーザーを特定できない GET 直アクセスは `/pin` または `/pin/login` へ戻します。
+- POST では `targetUserId` を受け取り、現在その端末で PIN 入力対象になっているユーザーと一致する必要があります。
+- 入力メールアドレスは、その対象ユーザーに登録されているメールアドレスと完全一致する必要があります。
 - 未認証レスポンスでは reset URL や token 相当の情報を返しません。
 - PIN 初期化リンクは `APP_PUBLIC_ORIGIN` を基準に組み立て、メールでのみ配布します。
 - `SMTP_*` と `APP_PUBLIC_ORIGIN` が未設定の環境では、このエンドポイントは `503` を返します。
